requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
hvac==0.10.3
idna==2.9
multidict==4.7.6
mysqlclient==1.4.6
okta==0.0.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
slackclient==2.6.2
urllib3==1.25.9
yarl==1.4.2
